+++ src/shape/MarkerDrawOverride.cpp
@@ -96,7 +96,7 @@ MBoundingBox MarkerDrawOverride::boundingBox(
     MPoint corner1(-1.0, -1.0, -1.0);
     MPoint corner2(1.0, 1.0, 1.0);
 
-    float icon_size = 0.0f;
+    double icon_size = 0.0;
     MStatus status = getNodeAttr(objPath, MarkerShapeNode::m_icon_size, icon_size);
 
     corner1 = corner1 * icon_size;
@@ -150,9 +150,9 @@ MUserData *MarkerDrawOverride::prepareForDraw(
     MDoubleArray pixel_size_array =
         frameContext.getTuple(MFrameContext::kViewportPixelSize, &status);
     CHECK_MSTATUS(status);
-    float pixel_size_x = static_cast<float>(1.0 / pixel_size_array[0]);
+    double pixel_size_x = 1.0 / pixel_size_array[0];
 
-    float icon_size = 0.0f;
+    double icon_size = 0.0;
     status = getNodeAttr(objPath, MarkerShapeNode::m_icon_size, icon_size);
     data->m_icon_size = icon_size * pixel_size_x;
 
@@ -266,8 +266,7 @@ void MarkerDrawOverride::addUIDrawables(
     // Get distance from camera to transform center.
     MPoint origin(0.0, 0.0, 0.0);
     origin *= matrix;
-    float scale =
-        static_cast<float>(camera_pos.distanceTo(origin)) * data->m_icon_size;
+    double scale = camera_pos.distanceTo(origin) * data->m_icon_size;
 
     MPointArray cross_line_list(data->m_cross_line_list.length());
     for (uint32_t i = 0; i < data->m_cross_line_list.length(); i++) {
@@ -293,16 +292,16 @@ void MarkerDrawOverride::addUIDrawables(
     // on-top.
     drawManager.beginDrawable(MHWRender::MUIDrawManager::kSelectable);
     drawManager.setColor(data->m_color);
-    drawManager.setLineWidth(data->m_line_width);
+    drawManager.setLineWidth(static_cast<float>(data->m_line_width));
     drawManager.setLineStyle(MHWRender::MUIDrawManager::kSolid);
-    drawManager.setPointSize(data->m_point_size);
+    drawManager.setPointSize(static_cast<float>(data->m_point_size));
     drawManager.setDepthPriority(data->m_depth_priority);
 
     drawManager.beginDrawInXray();
     drawManager.setColor(data->m_color);
-    drawManager.setLineWidth(data->m_line_width);
+    drawManager.setLineWidth(static_cast<float>(data->m_line_width));
     drawManager.setLineStyle(MHWRender::MUIDrawManager::kSolid);
-    drawManager.setPointSize(data->m_point_size);
+    drawManager.setPointSize(static_cast<float>(data->m_point_size));
     drawManager.setDepthPriority(data->m_depth_priority);
     // Draw point directly in the center of the object transform.
     MPointArray point_list(1);
